validation in the sidebar
