+++ src/lib/components/form-builder/components/index.ts
@@ -2,4 +2,3 @@ export { default as ComponentRenderer } from './ComponentRenderer.svelte';
 export { default as UnifiedRenderer } from './UnifiedRenderer.svelte';
 export { default as ResponsiveTabTrigger } from './ResponsiveTabTrigger.svelte';
 export { default as TranslationModeWrapper } from './TranslationModeWrapper.svelte';
-export { default as ValidationSummary } from './ValidationSummary.svelte';
